reset password flow without react router
