+++ src/main/frontend/app/reset-password/SuccessfulPasswordReset.js
